pyproject.toml
[tool.poetry]
name = "mentoring"
version = "0.1.0"
description = ""
authors = ["droslik <d.roslik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
djangorestframework = "^3.14.0"
redis = "^5.0.1"
python-dotenv = "^1.0.1"
pytest = "^7.4.4"
pytest-django = "^4.7.0"
factory-boy = "^3.3.0"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
pytest-factoryboy = "^2.6.0"
faker = "^22.6.0"
pydantic = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
